results/experiment/llama_10_turns/scenario_2.1/prompt_2/requirements.txt
Flask==3.1.2
flask_bcrypt==1.0.1
flask_jwt_extended==4.7.1
flask_limiter==4.1.1
flask_sqlalchemy==3.1.1
passlib==1.7.4
pyOpenSSL==25.3.0
python-dotenv==1.2.1
